add CacheService interface
